+++ package.json
@@ -14,8 +14,9 @@
     "react-dom": "^19.0.0"
   },
   "devDependencies": {
-    "@types/node": "22.14.0",
-    "@types/react": "19.1.0",
-    "typescript": "5.8.3"
+    "@types/node": "^22.14.0",
+    "@types/react": "^19.1.0",
+    "@types/react-dom": "^19.1.7",
+    "typescript": "^5.9.2"
   }
 }
